==> src/PerturbNet/03_get_cell_type_embedding.ipynb <==
import numpy as np
import pandas as pd
import torch
from tqdm import tqdm

df_rna = pd.read_csv('../../results/cell_type_embedding/rna_pca.csv', index_col=0)
df_rna['cell_type'] = df_rna.index
df_rna.reset_index(drop=True, inplace=True)
df_rna.head()

for cell_type in ['nk', 't_cd4', 't_cd8', 't_reg']:
    print(cell_type)
    
    df_train = pd.read_csv(f'../../results/PerturbNet/splited_data/train_{cell_type}.csv')
    df_valid = pd.read_csv(f'../../results/PerturbNet/splited_data/valid_{cell_type}.csv')
    
    df = pd.merge(df_train, df_rna, on='cell_type', how='left')
    
    train_x = df.iloc[:, 5:].values
    
    df = pd.merge(df_valid, df_rna, on='cell_type', how='left')
    valid_x = df.iloc[:, 5:].values
    
    np.savez(f'../../results/PerturbNet/cell_type_embedding_rna/train_{cell_type}.npz', x=train_x)
    np.savez(f'../../results/PerturbNet/cell_type_embedding_rna/valid_{cell_type}.npz', x=valid_x)

df_test = pd.read_csv(f'../../results/PerturbNet/splited_data/test.csv')
df = pd.merge(df_test, df_rna, on='cell_type', how='left')

test_x = df.iloc[:, 5:].values
np.savez(f'../../results/PerturbNet/cell_type_embedding_rna/test.npz', x=test_x)

df_atac = pd.read_csv('../../results/cell_type_embedding/atac_nmf.csv', index_col=0)
df_atac['cell_type'] = df_atac.index
df_atac.reset_index(drop=True, inplace=True)
df_atac.head()

for cell_type in ['nk', 't_cd4', 't_cd8', 't_reg']:
    print(cell_type)
    
    df_train = pd.read_csv(f'../../results/PerturbNet/splited_data/train_{cell_type}.csv')
    df_valid = pd.read_csv(f'../../results/PerturbNet/splited_data/valid_{cell_type}.csv')
    
    df = pd.merge(df_train, df_atac, on='cell_type', how='left')
    
    train_x = df.iloc[:, 5:].values
    
    df = pd.merge(df_valid, df_atac, on='cell_type', how='left')
    valid_x = df.iloc[:, 5:].values
    
    np.savez(f'../../results/PerturbNet/cell_type_embedding_atac/train_{cell_type}.npz', x=train_x)
    np.savez(f'../../results/PerturbNet/cell_type_embedding_atac/valid_{cell_type}.npz', x=valid_x)

df_test = pd.read_csv(f'../../results/PerturbNet/splited_data/test.csv')
df = pd.merge(df_test, df_rna, on='cell_type', how='left')

test_x = df.iloc[:, 5:].values
np.savez(f'../../results/PerturbNet/cell_type_embedding_atac/test.npz', x=test_x)

# df_embedding = df_rna.merge(df_atac)

# df_embedding.set_index('cell_type', inplace=True)
